Hold off on making coffeescript api.imply('modules').
If you're using 'coffeescript' but not 'ecmascript' or 'modules', you
should be able to keep doing that, and you can always `meteor add modules`
at any time. There's no need for the CoffeeScript plugin to force that
decision on you, and doing so would pose at least some risk for current
pure-CoffeeScript users trying to update to Meteor 1.3.3.