Fix misuse of Map object in Resolver.prototype.resolve.
Using this._resolveCache as an object rather than using the Map methods
(.has, .get, .set) almost works, miraculously, except of course when
resolving a module identifier whose name is "has", "get", or "set",
because then the cache appears to contain a matching entry, but the cached
value is a function object, not a proper byParentDir object.

Fixes #7215.