Symlink .meteor/local/plugin-cache in temporary test directory.
Because the cache directories are shared instead of copied, the results
now have a chance of surviving multiple invocations of `meteor test`.

This significantly reduces rebuild times when testing a large app; for
example, the meteor/tools/tests/apps/modules app now takes less than 10
seconds to rebuild (after restarting the process, not after a file change)
instead of 60 seconds or more.