Provide correct mup options for various Meteor versions
Meteor 1.4 does not work with `mupx` without forking it and providing your own docker image.  A better option is to use the kadirahq/mup developement version (also named `mup`).  A dockerImage is still necessary but hopefully the project will provide an option.  A work around is available in meteor/meteor#7475 but I'd like to avoid suggesting it publically at the moment.

Closes meteor/guide#529