Temporarily remove `cordova-support` isopacket
As it depends on spacebars and you can't have isopackets that depend
on non-checked out packages.

We'll need to solve this by either:
  1. Using submodules for third party isopacket dependencies
  2. Allowing isopackets to use troposphere deps
  3. Rewriting boilerplate generator to not use spacebars

Committing this so I can test a few more things and get a feel for
other problems that come up to help decide which solution is best.