Remove cordova-app-hello-world from dev bundle.
This package is depended upon by `cordova-lib`, and npm@3 hoists it to the
top-level dev_bundle/lib/node_modules directory.

We don't use this example app in Meteor, so we don't need it in the dev
bundle. If it is ever needed, developers can run

  meteor npm install -g cordova-app-hello-world

To include it in the dev bundle.

It would be really great if we could remove cordova-lib's extra copy of
npm, but that will probably have to wait until they update to npm@3.