Don't double-wrap callbacks.
When using Promises, a callback passed to .then() will sometimes pass through multiple .then() calls. If we aren't careful, we may end up wrapping a callback multiple times, resulting in a callback that creates a fiber which immediately creates another fiber and so on.

This is especially bad because in this chain situation of fibers-creating-fibers, none of the fibers complete before others are created, thus they force expansion of the fiber pool.

This, in turn, runs into a v8 bug in which each fiber created makes fiber-switching permanently slower, due to use of a linked list. See:

    https://github.com/meteor/meteor/issues/7747

This change simply adds a field to the wrapped callbacks that marks them as not needing further wrapping.

Fixes #10