Optimize meteorNpm.getProdPackageNames by not calling runNpmCommand.
For the galaxy-server application, this reduces the time taken to
determine production node_modules from 1400ms to 120ms.

As another optimization, we could record the hashes of all the
package.json files encountered in this scan, and return the same set of
production package names if those hashes have not changed, though the
benefits may not be worth the additional complexity.