Add crossorigin="anonymous" to script tags
If you serve your JS through a CDN, then most errors will report out to Kadira (or a similar tracking tool) as "Script error" because the JS file is hosted on another domain. Adding crossorigin="anonymous" to each script tag enables CORS (as long as the CDN is properly configured and the Meteor app also sets the Access-Control-Allow-Origin headers on JS files).