Explicit error in modules-runtime when loading client file on server
While following the tutorial this issue occurred:

https://github.com/meteor/tutorials/issues/45#issuecomment-222033814

Consider the knowledge of the user: This is his first attempt to work
with Meteor and imports.

This issue is a developer-error: Loading a client side intended file on
the server.

The normal error you will receive is:

Error: Cannot find module './body.html'

This proposal will show the following error:

Cannot find module './body.html' from /imports/ui/body.js on the
server. Are you sure this file is intended to load on the server?

The logical thing for the developer is to check their import, see if
the file exists etc. but because you have 2 main.js files it’s very
easy to make the error to load it in the server main.js file.

To prevent this issue and make it easier to find the fix I propose to
show more explicit errors to help the developer find the issue quicker.

My proposition has just 2 scenarios currently, it only detects paths
with:

/client
/imports/ui

Off course more specific errors may be introduced when needed.

I expect there is no performance penalty with this because Meteor is
already throwing an error and Meteor will crash anyway.

This change will not block expert users to do custom things as this
error will crash Meteor. It only adds more explicit error messages to
the console.

There are no tests included in the modules-runtime package. If we want
this error to be tested I need some direction in where to put tests for
this error message.