Set mongoOptions.server.reconnectTries = Infinity.
Setting mongoOptions.server.auto_reconnect was removed by #7880 via commit
0ffb9ac8247c53b9ab5658d75139b64771fd4de3, though it seems the Server
options still respect autoReconnect, even in version 2.2 of the driver:
http://mongodb.github.io/node-mongodb-native/2.2/api/Server.html

That said, having inspected the code of the `mongodb` package, I do not
believe this change is really critical, since the default value for
autoReconnect appears to be true.

More importantly, I can't find any support in the code of the `mongodb`
npm package or its dependencies for the claim that setting
mongoOptions.server.reconnectTries to 0 is the same as making it infinite,
so this commit sets it to Infinity.