Mark JS hot-code-push test "slow" so CircleCI will skip it.
This test has been the primary source of test suite unreliability in
recent weeks. By marking it "slow" I'm effectively disabling it on
CircleCI, which I think is acceptable because (1) it is, in fact, quite
slow; (2) it passes locally on Linux; and (3) it depends on the timing of
file change notifications, which is a recipe for test flakiness due to the
general non-determinism of file watching, but not necessarily a sign of a
broken developer experience.

Like other "slow" tests, I will keep running this test locally via

  meteor self-test --slow 'javascript hot code push'

as part of the manual QA testing that we do.