Throw error when trying to render missing `apibox` calls
With Meteor 1.4, the PackageAPI was decomposed into different packages in https://github.com/meteor/meteor/commit/14ee8775c3b2b0e6a90f911a68a1895110c6cf62.  This broke the existing apibox calls for these since they need their @memberof changed to `PackageNamespace` (same as was done with `PackageAPI`).

Without this change, an error was still produced by the template renderer (the same as seen in meteor/docs#56).  Unfortunately, the error didn't make it clear WHAT in particular failed, outputting at the very end withoug the specific `apidoc` call.

This commit changes the logic to throw an error for the actual missing apidoc call, which if not fixed, will cause the deployment to appear to finish but not actually generate the docs, as seen here:

https://circleci.com/gh/meteor/docs/223

This, combined with meteor/hexo-s3-deploy#2 should cause the deploy to fail properly.  I barely noticed this otherwise.