Remove function-name transform.
This transform takes the liberty of renaming local variables so they
don't collide with the names of functions and function-valued
properties, which doesn't seem even remotely safe to me, especially
considering that function.name is merely a convenience for several other
ways of getting the name of the current function, and has not proven to
be a widely used ECMAScript feature.

This transform can be added back via .babelrc if desired, as long as you
don't mind your variables being renamed.