Raise `emscripten` `TOTAL_MEMORY` to satisfy more complex dep-trees
This is an attempt to fix the issue reported in meteor/meteor#7533 which produces a `MINISAT-*: Cannot enlarge memory arrays` error.

I was able to reliably reproduce this using the repo provied by @mitar:
https://github.com/meteor/meteor/issues/7533#issuecomment-251893652

See my comments while reviewing this repo here:
https://github.com/meteor/meteor/issues/7533#issuecomment-258084308

In the case of this repo, it seems to be a case of a conflicting constraint, most likely bouncing back and forth a bit before it runs out of allocated memory.

In other cases, some users have reported that it will simply work after additional executions (of `meteor add <pkg>`, etc.).  I'd speculate that during calculations, MINISAT occasionally takes the "hard route" and runs out of memory before it finds the solution whereas sometimes it gets it right quickly - a matter of chance though.

This fixes the reproduction I had to work with.  There may be a bigger-picture solution here which will find the correct answer more efficiently while maintaining the lower memory footprint, but this 64MB configuration has worked well for a very, very long time and 128MB may as well.  All things considered, the amount of package data thrown at MINISAT is tremendous – sometimes using the entirety of the package catalog.