Make CI test runner more defensive and a bit more flexible.
This changes the CI script to make sure it runs all tests, regardless of the environment configuration while still allowing flexibility in the size of the container parallelism.

While not having any effect on the Meteor configuration this will now run all 8 test groups across a fewer number of containers, if necessary.

This fixes issues where false-positives are achieved on contributor PRs, like seen on meteor/meteor#7963.  I explained here:

https://github.com/meteor/meteor/pull/7963#issuecomment-264188162

This change will also allow the script to run tests in their entirety if the parallelism configuration is missing, such as when run from a local dev box.