Restore state of modules-runtime@0.7.8 as it was published
This code was reverted in 32140c87076a7533441c39d8a7b856bbbd983335 as part of a larger revert, but this particular change should actually remain since it is how modules-runtime@0.7.8 was and is published.  I had initially thought that this commit would also be rolled back (and result in a 0.7.9) but that did not yield the desired result.