Upgrade Facebook Graph API to use v2.8
In testing for #7715, I discovered that the v2.2 Graph API endpoint was still in use in the `facebook` package which was due to sunset on 2017-03-25.

See Facebook Graph API Changelog here:
  https://developers.facebook.com/docs/apps/changelog

When a Graph API endpoint is sunset, it (is claimed) to automatically turn over to the next more recent version, in this case v2.3.

v2.3 has a breaking-change over v2.2, notably listed in "Changes from v2.2 to v2.3":

> [Oauth Access Token] Format - The response format of https://www.facebook.com/v2.3/oauth/access_token returned when you exchange a code for an access_token now return valid JSON instead of being URL encoded. The new format of this response is {"access_token": {TOKEN}, "token_type":{TYPE}, "expires_in":{TIME}}. We made this update to be compliant with section 5.1 of RFC 6749.

This change updates both Graph APIs to v2.8 which has LTS until "At least October 2018".