Revert "Upgrade the `npm` npm package to version 4.1.2."
This reverts commit be0c8c3ee03c7d402848303dfe141f98cf98bfdd.

Although we thought the upgrade from npm 3.10.9 to 4.1.2 was worthwhile
and safe, this breaking change proved problematic:
https://github.com/npm/npm/blob/latest/CHANGELOG.md#no-more-partial-shrinkwraps-breaking

Specifically, if a Meteor package uses `Npm.depends`, and does not yet
have an `.npm/package/npm-shrinkwrap.json` file, Meteor will create a
partial shrinkwrap file in order to install npm dependencies, but (with
the changes in npm@4) transitive dependencies of the package will no
longer be installed.

Upgrading npm to a new major version was probably too much of a change for
a 1.4.2.x release, anyway, so we're reverting it for 1.4.2.7.

In case you can't wait for 1.4.2.7, you can "fix" this problem for
previous versions of Meteor by running

  meteor npm install --global npm@3.10.9

You can test that this downgrade worked by running

  meteor npm version