Update the 1.4.2 release notes to clarify Cordova `buildNumber` changes.
Cordova was updated to 6.2.0 in Meteor 1.4.2 and Cordova (itself) changed the way the `buildNumber` was generated.  Some users experienced problems with the various app stores having version number clashes due to this change but Cordova insists it was mandatory.

This was brought up in a discussion on https://github.com/meteor/meteor/issues/7205.

Resolves https://github.com/meteor/meteor/issues/7205