Change the `test-in-console` runner to use the Selenium ChromeDriver.
Previously, this process was using a `phantomjs` script which is likely
falling victim to its 1500+ unresolved issues.  I was lured into fixing
this because of failing tests which cropped up after bringing the Blaze
repository up to date.

This commit includes a runner which was partially borrowed from the
Blaze repository.