Remove version constraint from minifier-css in standard-minifier-css.
This restriction was preventing the standard-minifier-css from getting
the latest updates to minifier-css.  Typically, when publishing
a package outside of a release cycle, I believe we need to manually add
this constraint since `meteor publish-release` normally takes care of it
automatically, however in most cases, I don't believe this should be
(permanently) committed.

I may stand to be corrected, but will find out soon!