Remove versionsById IndexedDB object store.
Keeping track of the latest version for each module identifier was more
important when we were using localStorage, because it helped with
periodically removing unused module sources.

Since the storage limits for IndexedDB are substantially higher than for
localStorage, we can safely keep old versions in the cache, and reap the
benefits if those versions are ever needed again.

Eliminating this extra logic simplifies the caching logic considerably,
and also improves cache read performance.