Use 'minifier-babili' branch of 'blaze' repository.
This points the `blaze` submodule in the Meteor repository to:

https://github.com/meteor/blaze/tree/minifier-babili

This `minifier-babili` branch excludes new HTML5 tests which are failing
on PhantomJS, namely, the HTML5 tests which were added by:

https://github.com/meteor/blaze/commit/04db0d40cef726cac73736629c164d404a39d54d

While I believe the `VIDEO` tests are the most problematic, there is
nothing about the tests themselves which are lacking.  In fact, the
tests do run successfully on the Blaze repository and the changes have
actually been released for a number of months on Blaze itself, however
the submodule in Meteor hasn't been updated until now.

Meteor only uses `blaze` for generating templates in the
`boilerplate-generator` package and doesn't utilize the HTML5 in
question, nor is Meteor responsible for Blaze testing anymore.

Ultimately, our `test-in-console` needs to be updated to use
another option aside from PhantomJS, but I don't believe that should
hold up releasing this Babili minifier which solves a number of issues
in meteor/meteor#8378.